requirements.txt
djangorestframework
psycopg2-binary
requests
pdfkit 
qrcode
python-dotenv